Refactor API to Vercel serverless handlers
