+++ apps/api/api/health.ts
@@ -1,0 +1,5 @@
+import type { VercelRequest, VercelResponse } from '@vercel/node';
+
+export default function handler(_req: VercelRequest, res: VercelResponse) {
+  res.status(200).json({ status: 'ok' });
+}
