+++ apps/api/package.json
@@ -3,14 +3,11 @@
   "private": true,
   "version": "0.1.0",
   "type": "module",
-  "main": "server/index.js",
-  "scripts": {
-    "dev": "node --watch server/index.js",
-    "start": "node server/index.js"
-  },
+  "scripts": {},
   "dependencies": {
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
-    "express": "^4.19.2"
+    "dotenv": "^16.4.5"
+  },
+  "devDependencies": {
+    "@vercel/node": "^3.0.20"
   }
 }
